Add Feeds patch for resolving PHP errors.
This patch alters the Feeds module to resolve a bug related to
its integration with the way date_ical module passes
FeedsDateTime objects. This patch eliminates PHP errors related
to passing an object to a function expecting a string. At a high
level, this patch enables the ability to retrieve and map dates
from iCal feeds to Drupal date fields when using an iCal feed as
a Feeds source.